Fix genotype count to be correct.
Genotype parsing now happens in 2 passes. The first pass determines the
genotype count (for use with FORMAT fields with Number=G). The second
pass parses all the sample data with correct arity. The 1st pass will
only happen if both the GT field is present and a FORMAT field actually
requires it. Otherwise, no genotype count information is present and a
FORMAT field requiring it will cause the parser to fail.