Cache parsers based on metadata, alleles, and ploidy.
This adds a cache for parses generated from metadata, such as INFO and
FORMAT fields in a VCF file. This cache will hold onto such parsers
strongly, but the fact is that most VCF files will have very few unique
combinations of Metadata, alternate allelecount, and ploidy, so not much
memory will be wasted.