Sketch out lazy parsing of VCF sample data.
This adds a rough sketch of how we could handle lazily parsing the
sample data from a row of VCF data. The variants would always be parsed,
but parsing all the sample data would be delayed until it was
specifically requested. This would be meant to work with patches, so
patches using only variant data would be quicker.